requirements.txt
anyio==4.9.0

asyncpg==0.30.0
certifi==2025.7.14
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dotenv==1.1.1
python-telegram-bot==22.3
sniffio==1.3.1
SQLAlchemy==2.0.42
typing_extensions==4.14.1
uvicorn
fastapi
yookassa